* commit 'c3cf5b587a63c0d2f5987d6d5cf9397e4ce18dbc':
  REL-327179 Fix build of system tests
  REL-327179 Implement JobHistoryHandler and JobHistoryUpdater
  REL-325028 Add lock statement in ImportJob
  REL-325028 Changes after review
  REL-325028 Changes after review
  REL-325028 Changes after review
  REL-325028 Remove aliases
  REL-325028 Changes after review
  REL-325028 Release semaphore in ImportJob on both Completed and FatalException events; Set NativeFilePathSourceFieldName only if copy file option is set
  REL-325028 Add more UTs
  REL-325028 Add DestinationWorkspaceTagsCreationConfigurationTests
  REL-325028 Add SynchronizationConfigurationTests
  Fixed integration tests
  REL-325028 Fix SynchronizationExecutorTests
  REL-325028 Fix UTs
  REL-325028 Fix ConfigurationTests
  REL-325028 Fix ConfigurationStub
  REL-325028 Fix configuration issues
  REL-325028 Refactor ImportAPI instantiation
  REL-325028 Add ImportApiFactory

+++ Source/Relativity.Sync/Storage/DataSourceSnapshotConfiguration.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Relativity.Sync.Configuration;
-using Relativity.Sync.Transfer;
 
 namespace Relativity.Sync.Storage
 {
@@ -36,7 +35,7 @@ namespace Relativity.Sync.Storage
 		
 		public bool IsSnapshotCreated => !string.IsNullOrWhiteSpace(_cache.GetFieldValue<string>(SnapshotIdGuid));
 
-		public async Task SetSnapshotDataAsync(Guid runId, long totalRecordsCount)
+		public async Task SetSnapshotDataAsync(Guid runId, int totalRecordsCount)
 		{
 			await _cache.UpdateFieldValueAsync(SnapshotIdGuid, runId.ToString()).ConfigureAwait(false);
 			await _cache.UpdateFieldValueAsync(SnapshotRecordsCountGuid, totalRecordsCount).ConfigureAwait(false);
